scraping/requirements.txt
anyio==3.6.2
appdirs==1.4.4
async-generator==1.10
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
blinker==1.5
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
configparser==5.3.0
crayons==0.4.0
cryptography==39.0.2
cssselect==1.2.0
dateparser==1.1.7
demjson3==3.0.6
distro==1.8.0
dnspython==2.3.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.1.1
facebook-page-scraper==4.0.3
facebook-scraper==0.2.59
facebook-sdk==3.1.0
fake-useragent==1.1.1
fastapi==0.94.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
jsonschema==3.2.0
kaitaistruct==0.10
lxml==4.9.2
outcome==1.2.0
packaging==23.0
paramiko==3.1.0
parse==1.19.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.6
pydivert==2.1.0
pyee==8.2.2
pymongo==4.3.3
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305
PyYAML==5.4.1
regex==2022.10.31
requests==2.28.2
requests-html==0.10.0
selenium==4.1.0
selenium-wire==4.3.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
starlette==0.26.1
texttable==1.6.7
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
unicorn==2.0.1.post1
urllib3==1.26.15
urllib3-secure-extra==0.1.0
uvicorn==0.21.0
w3lib==2.1.1
webdriver-manager==3.2.2
websocket-client==0.59.0
websockets==10.4
wsproto==1.2.0
zipp==3.15.0
